Refaktoroi: Siirrä onr:ää käyttävä koodi omaan luokkaan;
ValpasOppijaService on jo tarpeeksi is luokka muutenkin, ja näin onr:ää suoraan
käyttävä toiminnallisuus on paremmin yhdessä luokassa.
